+++ client-go/applyconfiguration/core/v1beta1/internalconnectspec.go
@@ -13,6 +13,8 @@ import (
 // InternalConnectSpecApplyConfiguration represents a declarative configuration of the InternalConnectSpec type for use
 // with apply.
 type InternalConnectSpecApplyConfiguration struct {
+	Enabled                 *bool                                                  `json:"enabled,omitempty"`
+	Teardown                *bool                                                  `json:"teardown,omitempty"`
 	License                 *product.LicenseSpec                                   `json:"license,omitempty"`
 	Volume                  *product.VolumeSpec                                    `json:"volume,omitempty"`
 	NodeSelector            map[string]string                                      `json:"nodeSelector,omitempty"`
@@ -33,12 +35,29 @@ type InternalConnectSpecApplyConfiguration struct {
 	DatabaseSettings        *DatabaseSettingsApplyConfiguration                    `json:"databaseSettings,omitempty"`
 	ScheduleConcurrency     *int                                                   `json:"scheduleConcurrency,omitempty"`
 	AdditionalRuntimeImages []ConnectRuntimeImageSpecApplyConfiguration            `json:"additionalRuntimeImages,omitempty"`
+	AdditionalConfig        *string                                                `json:"additionalConfig,omitempty"`
 }
 
 // InternalConnectSpecApplyConfiguration constructs a declarative configuration of the InternalConnectSpec type for use with
 // apply.
 func InternalConnectSpec() *InternalConnectSpecApplyConfiguration {
 	return &InternalConnectSpecApplyConfiguration{}
+}
+
+// WithEnabled sets the Enabled field in the declarative configuration to the given value
+// and returns the receiver, so that objects can be built by chaining "With" function invocations.
+// If called multiple times, the Enabled field is set to the value of the last call.
+func (b *InternalConnectSpecApplyConfiguration) WithEnabled(value bool) *InternalConnectSpecApplyConfiguration {
+	b.Enabled = &value
+	return b
+}
+
+// WithTeardown sets the Teardown field in the declarative configuration to the given value
+// and returns the receiver, so that objects can be built by chaining "With" function invocations.
+// If called multiple times, the Teardown field is set to the value of the last call.
+func (b *InternalConnectSpecApplyConfiguration) WithTeardown(value bool) *InternalConnectSpecApplyConfiguration {
+	b.Teardown = &value
+	return b
 }
 
 // WithLicense sets the License field in the declarative configuration to the given value
@@ -215,5 +234,13 @@ func (b *InternalConnectSpecApplyConfiguration) WithAdditionalRuntimeImages(valu
 		}
 		b.AdditionalRuntimeImages = append(b.AdditionalRuntimeImages, *values[i])
 	}
+	return b
+}
+
+// WithAdditionalConfig sets the AdditionalConfig field in the declarative configuration to the given value
+// and returns the receiver, so that objects can be built by chaining "With" function invocations.
+// If called multiple times, the AdditionalConfig field is set to the value of the last call.
+func (b *InternalConnectSpecApplyConfiguration) WithAdditionalConfig(value string) *InternalConnectSpecApplyConfiguration {
+	b.AdditionalConfig = &value
 	return b
 }